# spec/openapi-3.1.yaml
openapi: 3.1.0
...
info:
  title: VedaCore Signals API
  version: "1.1.2"
  description: >-
    Authoritative OpenAPI 3.1 specification for the VedaCore Signals API.
    This file is the single source of truth for the REST API surface.
    The 3.0.x JSON at repository root is generated for SDKs.
  contact:
    name: VedaCore Support
    email: support@vedacore.io
  license:
    name: Proprietary
    url: https://vedacore.io/license
  x-security:
    adminEndpoints:
      - /stream/_resume
      - /stream/_topics
    requirements: "role=admin|owner OR scope includes 'stream:debug'"
    problem:
      status: 403
      code: FORBIDDEN_DEBUG
      title: Forbidden
servers:
  - url: https://api.vedacore.io
security:
  - HTTPBearer: []
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
    queryApiKey:
      type: apiKey
      in: query
      name: token
paths:
...
  /api/v1/kp/analysis:
    post:
      summary: KP Analysis (placeholder)
      tags: [ kp ]
      operationId: v1_kp_analysis
      responses:
        '200': { description: OK, content: { application/json: {} } }
        '401': { description: Unauthorized }
        '429': { description: Too Many Requests }